+++ backend/LabeledByAI.Services/LabeledByAI.Services.csproj
@@ -4,10 +4,12 @@
 		<TargetFramework>net8.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
+		<LangVersion>latest</LangVersion>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<PackageReference Include="Octokit.GraphQL" Version="0.4.0-beta" />
+		<PackageReference Include="Microsoft.Extensions.AI" Version="9.0.0-preview.9.24556.5" />
 	</ItemGroup>
 
 </Project>
