+++ backend/LabeledByAI/Program.cs
@@ -1,3 +1,4 @@
+using LabeledByAI.Services;
 using Microsoft.Azure.Functions.Worker.Builder;
 using Microsoft.Extensions.AI;
 using Microsoft.Extensions.DependencyInjection;
@@ -5,6 +6,9 @@ using Microsoft.Extensions.Hosting;
 using OpenAI;
 
 var builder = FunctionsApplication.CreateBuilder(args);
+
+builder.Services.AddSingleton<AIChatClient>();
+builder.Services.AddSingleton<GitHubBestLabelAIChatClient>();
 
 builder.AddServiceDefaults();
